moar fixes

+++ modules/store/src/helpers/fs.ts
@@ -4,7 +4,6 @@ import { EMPTY_STRINGIFIED_DATA } from "./constants";
 
 export const FILE_EXISTS = 1;
 export const FILE_DOESNT_EXIST = 0;
-
 export function fsRead(path: string): Promise<any> {
   return new Promise((resolve, reject) => {
     fs.readFile(path, (err, data) => {
@@ -15,7 +14,6 @@ export function fsRead(path: string): Promise<any> {
     });
   });
 }
-
 export function fsWrite(path: string, data: any): Promise<void> {
   return new Promise((resolve, reject) => {
     fs.writeFile(path, data, err => {
@@ -27,7 +25,10 @@ export function fsWrite(path: string, data: any): Promise<void> {
   });
 }
 
-export function fsUnlink(path: string): Promise<void> {
+export async function fsUnlink(path: string): Promise<void> {
+  if ((await checkFile(path)) === FILE_DOESNT_EXIST) {
+    return;
+  }
   return new Promise((resolve, reject) => {
     fs.unlink(path, err => {
       if (err) {
@@ -63,8 +64,7 @@ export function checkFile(path: string): Promise<number> {
 }
 
 export async function safeFsRead(path: string): Promise<any> {
-  const check = await checkFile(path);
-  if (check === FILE_DOESNT_EXIST) {
+  if ((await checkFile(path)) === FILE_DOESNT_EXIST) {
     const data = EMPTY_STRINGIFIED_DATA;
     await fsWrite(path, data);
     return data;
